Split filesystem client by responsibility
